+++ src/services/firebase/index.ts
@@ -1,5 +1,5 @@
 import { getFileURL } from './getFileURL';
-import { getSidebarCategories } from './getSidebarCategories';
+import { getSidebarCategories } from './getItems';
 import { verifyCode } from './verifyCode';
 
 export { getFileURL, getSidebarCategories, verifyCode };
